from #386 - I've fixed SPAT so that it works how I think it should... The track marked as "SPAT" should have nothing to do with queue or playback mode.
Playback should stop after the SPAT track is played, no matter how
playback ends up playing this track.  

If the SPAT track is in the playlist and
the playlist is on shuffle, playback could stop at any point.  

If the SPAT track is in the middle of the queue, the items at the 
end of the queue won't get played until playback is started again.

If shuffle is off and the SPAT track is three tracks forward in the playlist,
that's when playback will stop.