More gui tests for MainWindow
To do this I had to change event.py to not ever use asynchronous callbacks in
testing mode, as it's not needed.  To set testing mode, use event._TESTING =
True

I did the same thing in common.py so that @common.threading doesn't actually
return a threaded function, as this will also cause unexpected while
attempting to run things in a linear fashion

In track.py, _track_update_callback was calling "self[tag] = value", which was
causing _track_update_callback to be called which was calling "self[tag] =
value".... you see where this is going.  This is now fixed.