plugins/desktopcover: Adjust to new cover system.
Known bug: if the plugin is enabled while a song without cover is playing, the nocover image will be shown. This is hard to fix, because we only have access to the raw image data and pixbuf, so we don't know when the image is actually the nocover image.

A proper fix would be to have a "current cover" property in Exaile's main window, with appropriate getter/setter, that returns None when there is no cover.