Moodbar: repeat main setup for preview device
Repeat the pattern of setup from the main player moodbar for the preview
player: instantiates a global object for it, initialises the `player`
and `progress_bar` objects from the preview plugin object, instead of
the main one.

Then, because we don't know what order the plugins get setup, we check
to see if the previewdevice has already been activated. (If it hasn't we
can't import it, so catch that), and if so, use it to set up the preview
moodbar immediately.  If it hasn't bsen set up, then maybe it never
will,  in which case those callbacks never get called, or if it does
the callback should call the same setup method when the player gets
enabled.

A couple of places I forgot to change `player.PLAYER`
for `self.player`, so fixed those too.