Add "gi:" prefix support for plugin RequiredModules.
This allows plugins to state their dependency on GI modules. For example,
if a plugin requires WebKit2, add "gi:WebKit2" to RequiredModules.

The Gtk and Gst modules and their required dependencies (e.g. Pango) are
assumed to be present and should not be added to RequiredModules. This may
change in the future if Exaile switches to other frameworks.

Fixes https://github.com/exaile/exaile/issues/121