Make session detection more robust
On gnome 3.18+Wayland, DESKTOP_SESSION defaults to 'gnome-wayland', this is fixed by this commit.
XDG_CURRENT_SESSION is going to replace DESKTOP_SESSION, so add it to session detection.
Since XDG_CURRENT_SESSION is not very old (see [1]) keep DESKTOP_SESSION.

A more elegant way to solve this would be using Gtk.Application.inhibit(), then we could get rid of most of this code.
To do that Exaile would need to get ported to Gtk.Application first (much work included).

[1] https://askubuntu.com/questions/72549/how-to-determine-which-window-manager-is-running