Fix Export Files playlist dialog (#299)
The dialog to export files from a playlist (launched with the 'Export
Files' context menu option over such playlist tab) fails because it
tries to assign a value to a free variable in a closure (which are
read-only in Python).

This patch modifies the logic in the closure so the free variable is
only used as an R-value.