bpm: Fix autodetection from panel.
When the menu is opened from a panel, the playlist_view argument passed to
on_auto_menuitem (a menu.simple_menu_item() callback) is a panel.Panel
(which is not a widget). This caused playlist_view.get_toplevel() to fail.

This commit makes the code very specific: if playlist_view is a Widget, use
get_toplevel(); if it's a Panel, use parent; otherwise use None as the
dialog's parent window.

Fixes https://github.com/exaile/exaile/issues/296