panel notebook: Do not show context menu everywhere
Before this commit, the context menu of the notebook panel would not only be
shown on notebook tab, but also inside any panel when pressing the "menu" key.

After this commit, the context menu of the notebook panel is not shown any more
when pressing the "menu" key.

Opening the context menu by secondary mouse click is not affected by this commit,
since the code for that is in NotebookTab.

This is an attempt to fix #291.