Don't compare literal booleans using '=='
Don't compare literal booleans (True, False) using the equal operator,
use 'is' or use the value directly in the expression. These were
reported by pycodestyle ("E712 comparison to True should be 'if cond is
True:' or 'if cond:'", "E712 comparison to False should be 'if cond is
False:' or 'if not cond:'") by a search limited to the core source tree
(xl/ and xlgui/ directories).