playlist: massive speedup
Since all the data we need is stored in the track object in the PlaylistModel,
there's no reason to extract the information and store it in separate columns.
It turns out this is quite expensive. Instead, change the column renderers to
pull the text directly from the track, which has the added advantage of only
extracting data as needed, instead of rendering the entire playlist.