gst_utils: fixed parsing of track-number tag
When parsing track-number tag from the stream, be sure to convert
it from unsigned integer to string, which is assumed by the rest
of the player. While we are at it, if track-count is also available,
construct the double-number string.

This fixes the type error that was thrown when trying to display
properties of a track being played from the DAAP share.