panel/files: During enumerate_children, ignore errors on individual files.
When we do

    for info in gfile.enumerate_children(...):
        ...

internally PyGObject translates this to a bunch of C-style iteration calls.
On Windows, these can fail on particular files, e.g. on C:\*.sys , causing
the whole iteration to crumble. In other words, the Files panel currently
can't browse to C:\ .

In this patch we semi-manually wrap PyGObject's iteration, putting each
next() call in a try-except and skipping those that fail.

This should probably be moved to a utility module and used in other places
as well.