widgets/notebook: Tweak vertical notebook (i.e. panel) tabs a bit.
On Adwaita, the most important change is that this adds a bit of padding
between the label text and the bottom of the tab, which were previously too
close. This is slightly offset by reducing padding on the top side.

On other themes, this commit fixes countless issues. Now it looks pretty
good on all the themes I tested.