Playlist: prevent unselection only if multiple items are selected
Apply the "prevent unselection of all except the clicked item on
left clicks, required to preserve the selection for DnD" only if
there are actually multiple items selected. (As its function
appears to be preservation of multiple-item-selection for DnD).

Otherwise, this part of code appears to interfere with selection
of the (to-be-made) editable Comment column for currently selected
row.