playlist_columns: Don't delay the get_width call in on_width_changed.
When the delayed/batched function gets called, the widget may be in an
inconsistent state. On some older GTK+ versions (e.g. 3.18.9 on Linux Mint)
this resulted in column widths of 0 to be saved on startup.

This change calls get_width immediately in the signal handler and passes
the value to the delayed function.

Fixes: https://github.com/exaile/exaile/issues/580