py3: main.py: fixed handling of --all-data-dir argument
In python2, the alldatadir received via command-line was a byte
string, and thus we did a sequence of decode() and encode() to
remove any non-ASCII characters. This was necessary because if
there were UTF-8 characters in the string, the os.path.join()
call threw exception.

In python3, the alldatadir is a unicode string, so we need to
change the sequence to encode(), followed by decode(). It is also
worth noting that os.path.join() does not take issues with UTF-8
characters anymore, so perhaps the whole code block can be removed.

But for now, we replace the non-ASCII characters to ? to keep the
behavior consistent with the old version...