xlgui/widgets/playback: fixed signalling of marker being reached
We were comparing int(m.props.position * track_length) against
playback_time, but because the latter was a float, the comparison
was never true. So convert the playback_time to int before the
comparison to catch the moment when marker is reached.

Should fix the segment repeating in A-B repeat plugin.