plugins/bookmarks: reviewed futurize changes, fixed file load/save
In this plugin, the open() call in save function needs to be done
in non-binary mode, because otherwise json.dump() function throws
an error about trying to write an str instead of bytes.

The load function seems to work with both binary and
non-binary open(), so use non-binary to be consistent.