xl/collection: remove the use of basestring
In joiner(), we really need to distinguish only between lists
(which need to be joined) and other (which are returned as-is).

So rework the logic to use isinstance(value, list) instead of
isinstance(value, basestring). Also, always join using the
str-based separator, as tags should always be str objects now.