py3: xlgui/panel/collection: added comments about str() conversion
Added comments about the use of explicit str() conversion in places
where this is (or may be necessary).

While the sort tag value is usually a string, it is not always so;
for example, in case of track number tag.

The first_meaningful_char() is in practice always called with a
str argument due to the way our sort filters are currently
constructed. But keep the str() conversion there as well, in case
the sort groups get changed in the future.