py3: xlgui/panel/collection: removed the use of reduce()
The get_selected_tracks() was constructing list of track groups
(lists/tuples), merging them into a single list via reduce(),
then filtering out duplicates via set(), and finally converting
the set back to list.

This is now simplified by constructing the set on-the-fly, and
then converting it to the list.