py3: xl/trax/util: fixed error when sorting by non-string tags
When expanding the Collection tree, the following error is triggered:
TypeError: '<' not supported between instances of 'str' and 'int'

This is because Track.get_tag_sort() may return a non-string for
certain tags (track number). So for now, explicit conversion to
string is required.

Also fixed a typo in a comment at a similar explicit string conversion
in xlgui/panel/collection.py.