exaile_win: use error() to display exception information
Use error() to display exception information in the try/catch
block surrounding the exaile.main() call. The input() call in
the exception handling block was causing another exception
in situations when console is not available (i.e., pyinstaller
build). And this way, we can get the traceback displayed in
the error dialog, which makes things easier to debug.