Context: Also seek(0) when truncating the temporary stdout buffer
truncate(0) doesn't change the stream position. If we truncate without
seeking, the next time we write to the buffer it will start with nul
characters (and it keeps growing each time we write new things).

Then to GTK this looks like an empty string because it starts with nul.

I don't know why this didn't cause issues previously, as `truncate` behaves
the same in Python 2. Maybe PyGObject works differently there.

Fixes: https://github.com/exaile/exaile/issues/693