tray: Use Menu.popup_at_pointer instead of popup + StatusIcon.position_menu
Using popup + position_menu can cause the menu to show at the wrong place
in Windows. For example, if the taskbar is on the left (the tray at the
bottom left), the menu shows on the upper left corner, with most of the
items hidden until scrolled into.

popup_at_pointer requires GTK 3.22, which is already our specified minimum.