SeekProgressBar: Unset the color and font-size CSS properties
Adwaita blends the font color with the background and reduces the size,
making the text harder to read. This change restores the font color and
size to that of the parent element.

On some themes (e.g. BlueMenta, included with MATE) this can cause the
whole playback hbox to grow taller, but hopefully it still looks okay.

Fixes: https://github.com/exaile/exaile/issues/700