common: open_file_directory: Specify types on ctypes functions
Without this, ctypes assumes the return value of ILCreateFromPathW is a
signed int, whereas it is actually a pointer type (size_t, which is
unsigned on x86). When this gets passed to SHOpenFolderAndSelectItems,
Exaile segfaults. I don't know why the code used to work.

This fixes crash on Windows when pressing Open Directory on the Track
Properties dialog.