widgets/playlist: On dropping tracks, always scroll to them
This is a workaround for https://github.com/exaile/exaile/issues/685 ,
where DnD causes the view to scroll to the top. I can't figure out what
causes it, so let's go with a hack.

This change causes the view to scroll to the first of the dropped track,
putting it a quarter of the way from the top.

This makes the hidden `gui/scroll_when_appending_tracks` option (default:
False) obsolete because it's now the only supported behavior, so this
commit also removes that option. (The name of the option is misleading; it
only affects DnD.)

Fixes: https://github.com/exaile/exaile/issues/685
Closes: https://github.com/exaile/exaile/pull/684