xlgui: collection manager dialog: disallow addition of duplicated locations
The collection manager dialog disallows addition of locations that
are subdirectories of locations already existing in collection,
but does not disallow addition of location that is the same as
a location that already exists in collection.

This is because a Gio.File is not a prefix of itself, so we also
need to check for equality using Gio.File.equal().