CI: notify: Switch to using pydle's irccat, but comment it out for now
The extremely simple Bash IRC thing doesn't work from the CI server, and
was always flaky to begin with in my testing, probably because it writes
stuff without waiting for the server to reply.

This commit switches to using the irccat utility from pydle, which acts as
a proper IRC client. However, it's currently broken due to
https://github.com/Shizmob/pydle/issues/151

Until the fix makes it to a released version, I'm commenting out the whole
IRC notification job.