trax/util.is_valid_track: Change new_for_uri to new_for_commandline_arg
In Exaile core this is always called with a URI so it works, but the tests
call it with a local file path and we get test failures. (It must have
worked in earlier GIO versions, because our CI doesn't catch this.)

trax.Track uses new_for_commandline_arg for its location argument, so this
should be correct.

Fixes: https://github.com/exaile/exaile/issues/750