playlist: Remove no-op setting of __next_data.
This conditional block has no effect because self.__next_data will be
immediately overwritten at the end of the method. From this point in the
code there is no possibility of anything happening before that final
assignment because it will skip over the `else` block and we already
know that `next is not None` so it will skip over that block as well.

This is also the only place where self.__next_data is a tuple of length
2 when this class actually expects it to be None or a tuple of length 3.