Windows launcher: provide dummy sys.stderr and sys.stdout
Recent versions of PyInstaller do not provide dummy `sys.stderr`
and `sys.stdout` in noconsole/windowed mode anymore. Instead, they
are left at `None`, same as when script is ran using `pythonw.exe`.

Provide dummy file-like objects for `sys.stderr` and `sys.stdout`
to prevent errors in parts of the code that blindly assume that
the streams are available.