win installer: manually collect libsoup-2.4-1.dll and its dependencies
In GStreamer 1.22.1 used by our SDK 39, the `soup` GStreamer plugin,
which is used for radio streams, is not linked against the `soup`
shared library (DLL) anymore, so PyInstaller fails to pick up
the DLL and some of its dependencies. Failing to load the `soup`
plugin causes fall-back to `curl` one, which seems to fail to
load any radio stream.

Therefore, ensure that `libsoup-2.4-1.dll` is collected, by passing
it to PyInstaller via `binaries` in the spec file.