Fix .desktop & appdata generation
Previously this depended on the po/LINGUAS file created by the `pot`
target, potentially causing error on `make -j`.

This commit also changes the `locale` target to not depend on
`desktop_files` because `locale` is only meant to build the .mo files.

As a bonus, this also results in the .desktop and appdata files now not
being pointlessly created in the Windows CI build. We make the installation
of these files optional in the `make install` step.