lyricsmania: Fix unexisting element lookup
This is essentially a reversion of #861.

I received [a report](https://github.com/NixOS/nixpkgs/issues/442879)
that the lyricsmania plugin wasn't working and found that it wasn't
working for me either. When I dug into it, I found that it was caused
by a KeyError on the #video-musictory element, which I'm not seeing in
the browser either.

While it's possible that this element still exists in some situations,
it strikes me as more likely that the html has changed for good and
I'm just the first to notice and report. I could move this all to a
`try` block if we think it's necessary but I'm erring on the side of
deletion unless someone disagrees.

Also removing the unreachable `return` earlier in the file.